Changes behaviour of Total's percentile
Previously the percentile kept up with the current state of the Total timer. This was less than optimal, because in order to use the percentile as a metric of the quality of the run to this point, you had to watch and remember the value as you hit the split.

The new technique locks the percentile field to the percentile of the last split, thereby displaying the actual value of interest constantly.

This may also have fixed a bug with the handling skips and reverses in the percentile, but I can't be bothered to go back to see if it was actually a bug in the old behaviour, given that it's fixed with the new behaviour.