Modifies the total percentile behavior for the last split
Previously the last split would not update the total percentile at the end of the run. This is because widgets no longer know when a run is paused, so we can't guess when the run is done.
New behavior is to constantly update on the last split. This is a bit suboptimal IMHO, but it becomes more acceptable as the end of the run nears.
In future, a "Paused" value will be passed as part of a ReaderT pattern rewrite. At that time, we will estimate done-ness and return the behavior to its current behavior for other splits.