Rewrites the data format to include invalid splits
Due to split skipping, it is possible to have a personal best run which is not recorded. By recording all splits, rather than just valid ones, we can avoid this issue.
This commit updates the loader to support multiple versions of the data format (including backward compatibility and forward extensibility). All saves are (and will continue going forward to be) in the newest format.
This commit also adds a single-split validity 'or' key, 'i'. Pressing 'i' on a valid split will make it invalid, and vice versa. The behavior of 'skip' continues to invalidate the current and next splits.
Metrics which would be adversely affected by invalid splits (e.g., sumOfBests, levelBests, output) have had their input filtered appropriately.
All splits in earlier versions of the code (hereafter the "V1" data format) are considered to be valid when loaded.