Adds arbitrary depth split hierarchies
Hierarchies can now be established by indenting the split title by two spaces. Hierarchies of arbitrary depth can be achieved by repeated indentation. This has the added bonus of looking good in the interface.
Data format compatibility is maintained by including the path to the split in the split name recorded in the data file. This allows names in a hierarchy to be reused in another; however, names at any level of a given hierarchy (to include the top level) can not be duplicate.